+++ effect_subject/forms/mental_status_form.py
@@ -1,13 +1,20 @@
 from django import forms
+from edc_constants.constants import NOT_APPLICABLE
 from edc_crf.modelform_mixins import CrfModelFormMixin
 from edc_form_validators.form_validator import FormValidator
-from edc_glucose.form_validators import GlucoseFormValidatorMixin
+from edc_visit_schedule.utils import is_baseline
 
 from ..models import MentalStatus
 
 
-class MentalStatusFormValidator(GlucoseFormValidatorMixin, FormValidator):
-    pass
+class MentalStatusFormValidator(FormValidator):
+    def clean(self) -> None:
+        for reportable_field in ["reportable_as_ae", "patient_admitted"]:
+            self.applicable_if_true(
+                condition=not is_baseline(self.cleaned_data.get("subject_visit")),
+                not_applicable_msg="Expected 'Not applicable' at baseline.",
+                field_applicable=reportable_field,
+            )
 
 
 class MentalStatusForm(CrfModelFormMixin, forms.ModelForm):